move twitch and spotify to server lib
